scraper/requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.3
lxml==5.1.0
supabase==2.3.4
python-dotenv==1.0.0
openai==1.12.0
schedule==1.2.1
requests==2.31.0
python-dateutil==2.8.2
